refactor(execution-engine): extract task-runner-launch-support from task-runner
Moves launch/lease timing helpers (heartbeat interval, executor start timeout, startup-failure metadata, lease-expiry and SSH-retry helpers) out of task-runner into task-runner-launch-support, behavior unchanged.
One atomic slice of the step-4 large-file decomposition.
